Json

+++ pessoas.py
@@ -10,6 +10,13 @@ class Pessoa(Base):
 	nome = Column(String)
 	fl_professor = Column(Boolean)
 
+	def to_json(self):
+		return {
+			'id_pessoas': self.id_pessoas,
+			'nome': self.nome,
+			'fl_professor': self.fl_professor
+		}
+
 class AllPessoas():
 
 	def __init__(self, session):
@@ -19,20 +26,49 @@ class AllPessoas():
 		nova_pessoa = Pessoa()
 		nova_pessoa.nome = nome
 		nova_pessoa.fl_professor = fl_professor
-		self.session.add(nova_pessoa)
-		self.session.commit()
+		
+		try:
+			self.session.add(nova_pessoa)
+			self.session.commit()
+			return nova_pessoa.to_json()
+		except:
+			self.session.rollback()
+			raise
+
+	def readAll(self):
+		pessoas = self.session.query(Pessoa).all()
+		nova_lista = []
+		for pessoa in pessoas:
+			nova_lista.append(pessoa.to_json())
+		return nova_lista
 
 	def read(self, id):
 		pessoa = self.session.query(Pessoa).filter_by(id_pessoas = id).first()
-		return pessoa
+		return pessoa.to_json() if pessoa else None
 
 	def update(self, id_pessoas, nome, fl_professor):
-		pessoa = self.session.query(Pessoa).filter_by(id_pessoas = id_pessoas).first()
-		pessoa.nome = nome
-		pessoa.fl_professor = fl_professor
-		self.session.commit()
+		try:
+			pessoa = self.session.query(Pessoa).filter_by(id_pessoas = id_pessoas).first()
+			if not pessoa:
+				return None
+			pessoa.nome = nome
+			pessoa.fl_professor = fl_professor
+			self.session.commit()
+			return pessoa.to_json()
+		except:
+			self.session.rollback()
+			raise
 
 	def delete(self, id_pessoas):
-		pessoa = self.session.query(Pessoa).filter_by(id_pessoas = id_pessoas).first()
-		self.session.delete(pessoa)
-		self.session.commit()
+		try:
+			pessoa = self.session.query(Pessoa).filter_by(id_pessoas = id_pessoas).first()
+			if not pessoa:
+				return None
+			self.session.delete(pessoa)
+			self.session.commit()
+			return {
+				'mensagem': 'Pessoa apagada com sucesso!'
+			}
+		except:
+			self.session.rollback()
+			raise
